\name{ex01.15}
\alias{ex01.15}
\non_function{}
\title{data from exercise 1.15}
\description{
The \code{ex01.15} data frame has 40 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{ExamScor}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.15)
}
\keyword{datasets}
