\name{ex01.18}
\alias{ex01.18}
\non_function{}
\title{data from exercise 1.18}
\description{
The \code{ex01.18} data frame has 18 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Number.of.papers}{
a numeric vector
}
\item{Frequency}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.18)
}
\keyword{datasets}
