\name{ex01.19}
\alias{ex01.19}
\non_function{}
\title{data from exercise 1.19}
\description{
The \code{ex01.19} data frame has 15 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Number.of.particles}{
a numeric vector
}
\item{Frequency}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.19)
}
\keyword{datasets}
