\name{ex01.21}
\alias{ex01.21}
\non_function{}
\title{data from exercise 1.21}
\description{
The \code{ex01.21} data frame has 47 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{y}{
a numeric vector
}
\item{z}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.21)
}
\keyword{datasets}
