\name{ex01.24}
\alias{ex01.24}
\non_function{}
\title{data from exercise 1.24}
\description{
The \code{ex01.24} data frame has 100 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{ShearStr}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.24)
}
\keyword{datasets}
