\name{ex01.25}
\alias{ex01.25}
\non_function{}
\title{data from exercise 1.25}
\description{
The \code{ex01.25} data frame has 40 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{IDT}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.25)
}
\keyword{datasets}
