\name{ex01.27}
\alias{ex01.27}
\non_function{}
\title{data from exercise 1.27}
\description{
The \code{ex01.27} data frame has 50 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{solids}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.27)
}
\keyword{datasets}
