\name{ex01.28}
\alias{ex01.28}
\non_function{}
\title{data from exercise 1.28}
\description{
The \code{ex01.28} data frame has 48 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{radiation}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.28)
}
\keyword{datasets}
