\name{ex01.29}
\alias{ex01.29}
\non_function{}
\title{data from exercise 1.29}
\description{
The \code{ex01.29} data frame has 60 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{C5}{
a factor with levels
\code{B} 
\code{C} 
\code{F} 
\code{J} 
\code{M} 
\code{N} 
\code{O} 
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.29)
}
\keyword{datasets}
