\name{ex01.32}
\alias{ex01.32}
\non_function{}
\title{data from exercise 1.32}
\description{
The \code{ex01.32} data frame has 14 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Value}{
a numeric vector
}
\item{Cumulative.%}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(ex01.32)
}
\keyword{datasets}
