\name{ex01.37}
\alias{ex01.37}
\non_function{}
\title{data from exercise 1.37}
\description{
The \code{ex01.37} data frame has 11 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{C1}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.37)
}
\keyword{datasets}
