\name{ex01.38}
\alias{ex01.38}
\non_function{}
\title{data from exercise 1.38}
\description{
The \code{ex01.38} data frame has 9 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{pressure}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.38)
}
\keyword{datasets}
