\name{ex01.39}
\alias{ex01.39}
\non_function{}
\title{data from exercise 1.39}
\description{
The \code{ex01.39} data frame has 16 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{lives}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.39)
}
\keyword{datasets}
