\name{ex01.43}
\alias{ex01.43}
\non_function{}
\title{data from exercise 1.43}
\description{
The \code{ex01.43} data frame has 10 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Lifetime}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.43)
}
\keyword{datasets}
