\name{ex01.49}
\alias{ex01.49}
\non_function{}
\title{data from exercise 1.49}
\description{
The \code{ex01.49} data frame has 17 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{area}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.49)
}
\keyword{datasets}
