\name{ex01.50}
\alias{ex01.50}
\non_function{}
\title{data from exercise 1.50}
\description{
The \code{ex01.50} data frame has 7 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{LoadLife}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.50)
}
\keyword{datasets}
