\name{ex01.51}
\alias{ex01.51}
\non_function{}
\title{data from exercise 1.51}
\description{
The \code{ex01.51} data frame has 19 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{InducTm}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.51)
}
\keyword{datasets}
