\name{ex01.59}
\alias{ex01.59}
\non_function{}
\title{data from exercise 1.59}
\description{
The \code{ex01.59} data frame has 78 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{conc}{
a numeric vector
}
\item{cause}{
a factor with levels
\code{ED} 
\code{non-ED} 
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.59)
}
\keyword{datasets}
