\name{ex01.65}
\alias{ex01.65}
\non_function{}
\title{data from exercise 1.65}
\description{
The \code{ex01.65} data frame has 4 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{HC.(gm/mi)}{
a numeric vector
}
\item{CO.(gm/mi)}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.65)
}
\keyword{datasets}
