\name{ex01.67}
\alias{ex01.67}
\non_function{}
\title{data from exercise 1.67}
\description{
The \code{ex01.67} data frame has 15 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{CO.conc}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.67)
}
\keyword{datasets}
