\name{ex01.71}
\alias{ex01.71}
\non_function{}
\title{data from exercise 1.71}
\description{
The \code{ex01.71} data frame has 20 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{C1}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.71)
}
\keyword{datasets}
