\name{ex01.73}
\alias{ex01.73}
\non_function{}
\title{data from exercise 1.73}
\description{
The \code{ex01.73} data frame has 15 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Type.1}{
a numeric vector
}
\item{Type.2}{
a numeric vector
}
\item{Type.3}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.73)
}
\keyword{datasets}
