\name{ex01.75}
\alias{ex01.75}
\non_function{}
\title{data from exercise 1.75}
\description{
The \code{ex01.75} data frame has 45 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Time}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.75)
}
\keyword{datasets}
