\name{ex01.78}
\alias{ex01.78}
\non_function{}
\title{data from exercise 1.78}
\description{
The \code{ex01.78} data frame has 15 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Length}{
a factor with levels
\code{10-<12} 
\code{12-<14} 
\code{14-<16} 
\code{16-<18} 
\code{18-<20} 
\code{20-<22} 
\code{22-<24} 
\code{24-<26} 
\code{26-<28} 
\code{28-<30} 
\code{30-<35} 
\code{35-<40} 
\code{40-<45} 
\code{6-<8} 
\code{8-<10} 
}
\item{Frequency}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.78)
}
\keyword{datasets}
