\name{ex01.81}
\alias{ex01.81}
\non_function{}
\title{data from exercise 1.81}
\description{
The \code{ex01.81} data frame has 26 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{rainfall}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex01.81)
}
\keyword{datasets}
