\name{ex04.87}
\alias{ex04.87}
\non_function{}
\title{data from exercise 4.87}
\description{
The \code{ex04.87} data frame has 16 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{life.hr.}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex04.87)
}
\keyword{datasets}
