\name{ex06.03}
\alias{ex06.03}
\non_function{}
\title{data from exercise 6.3}
\description{
The \code{ex06.03} data frame has 16 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{C1}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex06.03)
}
\keyword{datasets}
