\name{ex06.06}
\alias{ex06.06}
\non_function{}
\title{data from exercise 6.6}
\description{
The \code{ex06.06} data frame has 31 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Strmflow}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex06.06)
}
\keyword{datasets}
