\name{ex06.09}
\alias{ex06.09}
\non_function{}
\title{data from exercise 6.9}
\description{
The \code{ex06.09} data frame has 150 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{scratch}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex06.09)
}
\keyword{datasets}
