\name{ex06.25}
\alias{ex06.25}
\non_function{}
\title{data from exercise 6.25}
\description{
The \code{ex06.25} data frame has 10 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{strength}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex06.25)
}
\keyword{datasets}
