\name{ex07.10}
\alias{ex07.10}
\non_function{}
\title{data from exercise 7.10}
\description{
The \code{ex07.10} data frame has 15 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{lifetime}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex07.10)
}
\keyword{datasets}
