\name{ex07.26}
\alias{ex07.26}
\non_function{}
\title{data from exercise 7.26}
\description{
The \code{ex07.26} data frame has 11 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Number.of.absences}{
a numeric vector
}
\item{Frequency}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex07.26)
}
\keyword{datasets}
