\name{ex07.43}
\alias{ex07.43}
\non_function{}
\title{data from exercise 7.43}
\description{
The \code{ex07.43} data frame has 22 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{toughnss}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex07.43)
}
\keyword{datasets}
