\name{ex07.44}
\alias{ex07.44}
\non_function{}
\title{data from exercise 7.44}
\description{
The \code{ex07.44} data frame has 15 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{C1}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex07.44)
}
\keyword{datasets}
