\name{ex07.47}
\alias{ex07.47}
\non_function{}
\title{data from exercise 7.47}
\description{
The \code{ex07.47} data frame has 18 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{ResidGas}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex07.47)
}
\keyword{datasets}
