\name{ex07.54}
\alias{ex07.54}
\non_function{}
\title{data from exercise 7.54}
\description{
The \code{ex07.54} data frame has 16 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{pul.comp}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex07.54)
}
\keyword{datasets}
