\name{ex08.32}
\alias{ex08.32}
\non_function{}
\title{data from exercise 8.32}
\description{
The \code{ex08.32} data frame has 12 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{C1}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex08.32)
}
\keyword{datasets}
