\name{ex08.53}
\alias{ex08.53}
\non_function{}
\title{data from exercise 8.53}
\description{
The \code{ex08.53} data frame has 13 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{times}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex08.53)
}
\keyword{datasets}
