\name{ex08.55}
\alias{ex08.55}
\non_function{}
\title{data from exercise 8.55}
\description{
The \code{ex08.55} data frame has 6 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{CO.conc}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex08.55)
}
\keyword{datasets}
