\name{ex08.64}
\alias{ex08.64}
\non_function{}
\title{data from exercise 8.64}
\description{
The \code{ex08.64} data frame has 8 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{SoilHeat}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex08.64)
}
\keyword{datasets}
