\name{ex08.68}
\alias{ex08.68}
\non_function{}
\title{data from exercise 8.68}
\description{
The \code{ex08.68} data frame has 20 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{time}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex08.68)
}
\keyword{datasets}
