\name{ex09.11}
\alias{ex09.11}
\non_function{}
\title{data from exercise 9.11}
\description{
The \code{ex09.11} data frame has 2 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Age.(days)}{
a numeric vector
}
\item{Sample.Size}{
a numeric vector
}
\item{Sample.Mean}{
a numeric vector
}
\item{Sample.Standard.Deviation}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.11)
}
\keyword{datasets}
