\name{ex09.23}
\alias{ex09.23}
\non_function{}
\title{data from exercise 9.23}
\description{
The \code{ex09.23} data frame has 32 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{extens}{
a numeric vector
}
\item{quality}{
a factor with levels
\code{H} 
\code{P} 
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.23)
}
\keyword{datasets}
