\name{ex09.27}
\alias{ex09.27}
\non_function{}
\title{data from exercise 9.27}
\description{
The \code{ex09.27} data frame has 2 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Type.of.Player}{
a factor with levels
\code{Advanced} 
\code{Intermediate} 
}
\item{Sample.size}{
a numeric vector
}
\item{Sample.mean}{
a numeric vector
}
\item{Sample.standard.deviation}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.27)
}
\keyword{datasets}
