\name{ex09.30}
\alias{ex09.30}
\non_function{}
\title{data from exercise 9.30}
\description{
The \code{ex09.30} data frame has 2 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Type}{
a factor with levels
\code{Commercial carbon grid} 
\code{Fiberglass grid} 
}
\item{Sample.size}{
a numeric vector
}
\item{Sample.mean}{
a numeric vector
}
\item{Sample.standard.deviation}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.30)
}
\keyword{datasets}
