\name{ex09.31}
\alias{ex09.31}
\non_function{}
\title{data from exercise 9.31}
\description{
The \code{ex09.31} data frame has 11 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{C1}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.31)
}
\keyword{datasets}
