\name{ex09.32}
\alias{ex09.32}
\non_function{}
\title{data from exercise 9.32}
\description{
The \code{ex09.32} data frame has 2 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Type.of.wood}{
a factor with levels
\code{Douglas fir} 
\code{Red oak} 
}
\item{Sample.size}{
a numeric vector
}
\item{Sample.mean}{
a numeric vector
}
\item{Sample.standard.deviation}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.32)
}
\keyword{datasets}
