\name{ex09.33}
\alias{ex09.33}
\non_function{}
\title{data from exercise 9.33}
\description{
The \code{ex09.33} data frame has 2 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Treatment}{
a factor with levels
\code{Control} 
\code{Steroid} 
}
\item{Sample.size}{
a numeric vector
}
\item{Sample.mean}{
a numeric vector
}
\item{Sample.standard.deviation}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.33)
}
\keyword{datasets}
