\name{ex09.37}
\alias{ex09.37}
\non_function{}
\title{data from exercise 9.37}
\description{
The \code{ex09.37} data frame has 33 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{House}{
a numeric vector
}
\item{Indoor}{
a numeric vector
}
\item{Outdoor}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.37)
}
\keyword{datasets}
