\name{ex09.38}
\alias{ex09.38}
\non_function{}
\title{data from exercise 9.38}
\description{
The \code{ex09.38} data frame has 15 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Test.condition}{
a numeric vector
}
\item{Normal}{
a numeric vector
}
\item{High}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.38)
}
\keyword{datasets}
