\name{ex09.39}
\alias{ex09.39}
\non_function{}
\title{data from exercise 9.39}
\description{
The \code{ex09.39} data frame has 14 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Infant}{
a numeric vector
}
\item{Isotopic.method}{
a numeric vector
}
\item{Test.weighing.method}{
a numeric vector
}
\item{Difference}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.39)
}
\keyword{datasets}
