\name{ex09.40}
\alias{ex09.40}
\non_function{}
\title{data from exercise 9.40}
\description{
The \code{ex09.40} data frame has 16 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{pipe}{
a numeric vector
}
\item{brush}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.40)
}
\keyword{datasets}
