\name{ex09.41}
\alias{ex09.41}
\non_function{}
\title{data from exercise 9.41}
\description{
The \code{ex09.41} data frame has 9 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{black}{
a numeric vector
}
\item{white}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.41)
}
\keyword{datasets}
