\name{ex09.43}
\alias{ex09.43}
\non_function{}
\title{data from exercise 9.43}
\description{
The \code{ex09.43} data frame has 15 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{C1}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.43)
}
\keyword{datasets}
