\name{ex09.44}
\alias{ex09.44}
\non_function{}
\title{data from exercise 9.44}
\usage{data(ex09.44)}
\description{
  The \code{ex09.44} data frame has 9 rows and 3 columns representing
  the results of an experiment to compare the yields (kg/ha) or Sundance
  winter wheat and Manitou spring wheat.  }
\format{
  This data frame contains the following columns:
  \describe{
    \item{Location}{a numeric vector indicating the location}
    \item{Sundance}{a numeric vector of yields (kg/ha) of Sundance
      winter wheat}
    \item{Manitou}{a numeric vector of yields (kg/ha) of Manitou spring wheat}
  }
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\references{
(1991) ``Agronomic performance of winter versus spring wheat'',
\emph{Agronomic Journal}, 527-531.
}
\examples{
data(ex09.44)
str(ex09.44)
}
\keyword{datasets}
