\name{ex09.63}
\alias{ex09.63}
\non_function{}
\title{data from exercise 9.63}
\description{
The \code{ex09.63} data frame has 4 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Epoxy}{
a numeric vector
}
\item{MMA}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.63)
}
\keyword{datasets}
