\name{ex09.65}
\alias{ex09.65}
\non_function{}
\title{data from exercise 9.65}
\description{
The \code{ex09.65} data frame has 2 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Method}{
a factor with levels
\code{Fixed} 
\code{Floating} 
}
\item{n}{
a numeric vector
}
\item{mean}{
a numeric vector
}
\item{SD}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.65)
}
\keyword{datasets}
