\name{ex09.66}
\alias{ex09.66}
\non_function{}
\title{data from exercise 9.66}
\description{
The \code{ex09.66} data frame has 8 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Fertilizer.plots}{
a numeric vector
}
\item{Control.plots}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.66)
}
\keyword{datasets}
