\name{ex09.70}
\alias{ex09.70}
\non_function{}
\title{data from exercise 9.70}
\description{
The \code{ex09.70} data frame has 2 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Type}{
a factor with levels
\code{With.side.coating} 
\code{Without.side.coating} 
}
\item{size}{
a numeric vector
}
\item{mean}{
a numeric vector
}
\item{SD}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.70)
}
\keyword{datasets}
