\name{ex09.72}
\alias{ex09.72}
\non_function{}
\title{data from exercise 9.72}
\description{
The \code{ex09.72} data frame has 17 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Motor}{
a numeric vector
}
\item{Commutator}{
a numeric vector
}
\item{Pinion}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.72)
}
\keyword{datasets}
