\name{ex09.76}
\alias{ex09.76}
\non_function{}
\title{data from exercise 9.76}
\description{
The \code{ex09.76} data frame has 2 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Site}{
a factor with levels
\code{Clean} 
\code{Steam.plant} 
}
\item{size}{
a numeric vector
}
\item{mean}{
a numeric vector
}
\item{SD}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.76)
}
\keyword{datasets}
