\name{ex09.77}
\alias{ex09.77}
\non_function{}
\title{data from exercise 9.77}
\description{
The \code{ex09.77} data frame has 5 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Twist}{
a numeric vector
}
\item{Control}{
a numeric vector
}
\item{Heated}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.77)
}
\keyword{datasets}
