\name{ex09.78}
\alias{ex09.78}
\non_function{}
\title{data from exercise 9.78}
\description{
The \code{ex09.78} data frame has 2 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Group}{
a factor with levels
\code{Elderly.men} 
\code{Young} 
}
\item{size}{
a numeric vector
}
\item{mean}{
a numeric vector
}
\item{stdError}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.78)
}
\keyword{datasets}
