\name{ex09.79}
\alias{ex09.79}
\non_function{}
\title{data from exercise 9.79}
\description{
The \code{ex09.79} data frame has 8 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Good.visibility}{
a numeric vector
}
\item{Poor.visibility}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.79)
}
\keyword{datasets}
