\name{ex09.84}
\alias{ex09.84}
\non_function{}
\title{data from exercise 9.84}
\description{
The \code{ex09.84} data frame has 4 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Treatment}{
a numeric vector
}
\item{n}{
a numeric vector
}
\item{SD}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.84)
}
\keyword{datasets}
