\name{ex09.86}
\alias{ex09.86}
\non_function{}
\title{data from exercise 9.86}
\description{
The \code{ex09.86} data frame has 8 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Carpet}{
a numeric vector
}
\item{NoCarpet}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.86)
}
\keyword{datasets}
