\name{ex09.90}
\alias{ex09.90}
\non_function{}
\title{data from exercise 9.90}
\description{
The \code{ex09.90} data frame has 8 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Number}{
a numeric vector
}
\item{Region1}{
a numeric vector
}
\item{Region2}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex09.90)
}
\keyword{datasets}
