\name{ex10.06}
\alias{ex10.06}
\non_function{}
\title{data from exercise 10.6}
\description{
The \code{ex10.06} data frame has 40 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{totalFe}{
a numeric vector
}
\item{type}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex10.06)
}
\keyword{datasets}
