\name{ex10.08}
\alias{ex10.08}
\non_function{}
\title{data from exercise 10.8}
\description{
The \code{ex10.08} data frame has 35 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{stiffnss}{
a numeric vector
}
\item{length}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex10.08)
}
\keyword{datasets}
