\name{ex10.09}
\alias{ex10.09}
\non_function{}
\title{data from exercise 10.9}
\description{
The \code{ex10.09} data frame has 6 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Wheat}{
a numeric vector
}
\item{Barley}{
a numeric vector
}
\item{Maize}{
a numeric vector
}
\item{Oats}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex10.09)
}
\keyword{datasets}
