\name{ex10.18}
\alias{ex10.18}
\non_function{}
\title{data from exercise 10.18}
\description{
The \code{ex10.18} data frame has 20 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{growth}{
a numeric vector
}
\item{hormone}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex10.18)
}
\keyword{datasets}
