\name{ex10.22}
\alias{ex10.22}
\non_function{}
\title{data from exercise 10.22}
\description{
The \code{ex10.22} data frame has 18 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{yield}{
a numeric vector
}
\item{EC}{
a numeric vector
}
\item{ECf}{
a factor with levels
\code{A} 
\code{B} 
\code{C} 
\code{D} 
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex10.22)
}
\keyword{datasets}
