\name{ex10.32}
\alias{ex10.32}
\non_function{}
\title{data from exercise 10.32}
\description{
The \code{ex10.32} data frame has 5 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{A}{
a numeric vector
}
\item{B}{
a numeric vector
}
\item{C}{
a numeric vector
}
\item{D}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex10.32)
}
\keyword{datasets}
