\name{ex10.36}
\alias{ex10.36}
\non_function{}
\title{data from exercise 10.36}
\description{
The \code{ex10.36} data frame has 4 rows and 5 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{L.D}{
a numeric vector
}
\item{R}{
a numeric vector
}
\item{R.L}{
a numeric vector
}
\item{C}{
a numeric vector
}
\item{C.L}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex10.36)
}
\keyword{datasets}
