\name{ex10.37}
\alias{ex10.37}
\non_function{}
\title{data from exercise 10.37}
\description{
The \code{ex10.37} data frame has 7 rows and 5 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Brand.1}{
a numeric vector
}
\item{Brand.2}{
a numeric vector
}
\item{Brand.3}{
a numeric vector
}
\item{Brand.4}{
a numeric vector
}
\item{Brand.5}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex10.37)
}
\keyword{datasets}
