\name{ex11.02}
\alias{ex11.02}
\non_function{}
\title{data from exercise 11.2}
\description{
The \code{ex11.02} data frame has 12 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{corrosn}{
a numeric vector
}
\item{coating}{
a numeric vector
}
\item{SoilType}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex11.02)
}
\keyword{datasets}
