\name{ex11.03}
\alias{ex11.03}
\non_function{}
\title{data from exercise 11.3}
\description{
The \code{ex11.03} data frame has 16 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{transfer}{
a numeric vector
}
\item{gas}{
a numeric vector
}
\item{liquid}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex11.03)
}
\keyword{datasets}
