\name{ex11.05}
\alias{ex11.05}
\non_function{}
\title{data from exercise 11.5}
\description{
The \code{ex11.05} data frame has 20 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{force}{
a numeric vector
}
\item{angle}{
a numeric vector
}
\item{connectr}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex11.05)
}
\keyword{datasets}
