\name{ex11.10}
\alias{ex11.10}
\non_function{}
\title{data from exercise 11.10}
\description{
The \code{ex11.10} data frame has 10 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Batch}{
a numeric vector
}
\item{Method.A}{
a numeric vector
}
\item{Method.B}{
a numeric vector
}
\item{Method.C}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(ex11.10)
}
\keyword{datasets}
