\name{ex11.15}
\alias{ex11.15}
\non_function{}
\title{data from exercise 11.15}
\description{
The \code{ex11.15} data frame has 18 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Sand}{
a numeric vector
}
\item{Carbon}{
a numeric vector
}
\item{Hardness}{
a numeric vector
}
\item{Strength}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex11.15)
}
\keyword{datasets}
