\name{ex11.16}
\alias{ex11.16}
\non_function{}
\title{data from exercise 11.16}
\description{
The \code{ex11.16} data frame has 18 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Response}{
a numeric vector
}
\item{Formulat}{
a numeric vector
}
\item{Speed}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex11.16)
}
\keyword{datasets}
