\name{ex11.17}
\alias{ex11.17}
\non_function{}
\title{data from exercise 11.17}
\description{
The \code{ex11.17} data frame has 18 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{acidity}{
a numeric vector
}
\item{Coal}{
a numeric vector
}
\item{NaOH}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex11.17)
}
\keyword{datasets}
