\name{ex11.18}
\alias{ex11.18}
\non_function{}
\title{data from exercise 11.18}
\description{
The \code{ex11.18} data frame has 18 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Yield}{
a numeric vector
}
\item{Speed}{
a numeric vector
}
\item{Formulation}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex11.18)
}
\keyword{datasets}
