\name{ex11.20}
\alias{ex11.20}
\non_function{}
\title{data from exercise 11.20}
\description{
The \code{ex11.20} data frame has 18 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{current}{
a numeric vector
}
\item{glass}{
a numeric vector
}
\item{phosphor}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex11.20)
}
\keyword{datasets}
