\name{ex11.29}
\alias{ex11.29}
\non_function{}
\title{data from exercise 11.29}
\description{
The \code{ex11.29} data frame has 96 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{length}{
a numeric vector
}
\item{time}{
a numeric vector
}
\item{heat}{
a numeric vector
}
\item{machine}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex11.29)
}
\keyword{datasets}
