\name{ex11.31}
\alias{ex11.31}
\non_function{}
\title{data from exercise 11.31}
\description{
The \code{ex11.31} data frame has 27 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Yield}{
a numeric vector
}
\item{time}{
a numeric vector
}
\item{tempture}{
a numeric vector
}
\item{pressure}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex11.31)
}
\keyword{datasets}
