\name{ex11.34}
\alias{ex11.34}
\non_function{}
\title{data from exercise 11.34}
\description{
The \code{ex11.34} data frame has 36 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Sales}{
a numeric vector
}
\item{store}{
a numeric vector
}
\item{week}{
a numeric vector
}
\item{shelf}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex11.34)
}
\keyword{datasets}
