\name{ex11.35}
\alias{ex11.35}
\non_function{}
\title{data from exercise 11.35}
\description{
The \code{ex11.35} data frame has 25 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Moisture}{
a numeric vector
}
\item{plant}{
a numeric vector
}
\item{leafsize}{
a numeric vector
}
\item{time}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex11.35)
}
\keyword{datasets}
