\name{ex11.39}
\alias{ex11.39}
\non_function{}
\title{data from exercise 11.39}
\description{
The \code{ex11.39} data frame has 24 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{cleaning}{
a numeric vector
}
\item{detergnt}{
a numeric vector
}
\item{carbonat}{
a numeric vector
}
\item{cellulos}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex11.39)
}
\keyword{datasets}
