\name{ex11.43}
\alias{ex11.43}
\non_function{}
\title{data from exercise 11.43}
\description{
The \code{ex11.43} data frame has 16 rows and 5 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{duration}{
a numeric vector
}
\item{vibratn}{
a numeric vector
}
\item{tempture}{
a numeric vector
}
\item{altitude}{
a numeric vector
}
\item{firing}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex11.43)
}
\keyword{datasets}
