\name{ex11.48}
\alias{ex11.48}
\non_function{}
\title{data from exercise 11.48}
\description{
The \code{ex11.48} data frame has 8 rows and 5 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{thrust}{
a numeric vector
}
\item{vibratn}{
a numeric vector
}
\item{tempture}{
a numeric vector
}
\item{altitude}{
a numeric vector
}
\item{firing}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex11.48)
}
\keyword{datasets}
