\name{ex11.55}
\alias{ex11.55}
\non_function{}
\title{data from exercise 11.55}
\description{
The \code{ex11.55} data frame has 16 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Extraction}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex11.55)
}
\keyword{datasets}
