\name{ex11.59}
\alias{ex11.59}
\non_function{}
\title{data from exercise 11.59}
\description{
The \code{ex11.59} data frame has 36 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Cure.Time.1}{
a numeric vector
}
\item{Adhesive.type}{
a factor with levels
\code{Copper} 
\code{Nickel} 
}
\item{Adhesive.factor}{
a numeric vector
}
\item{Cure.Time}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex11.59)
}
\keyword{datasets}
