\name{ex12.04}
\alias{ex12.04}
\non_function{}
\title{data from exercise 12.4}
\description{
The \code{ex12.04} data frame has 14 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{x}{
a numeric vector
}
\item{y}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex12.04)
}
\keyword{datasets}
