\name{ex12.19}
\alias{ex12.19}
\non_function{}
\title{data from exercise 12.19}
\description{
The \code{ex12.19} data frame has 14 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{area}{
a numeric vector
}
\item{emission}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex12.19)
}
\keyword{datasets}
