\name{ex12.20}
\alias{ex12.20}
\non_function{}
\title{data from exercise 12.20}
\description{
The \code{ex12.20} data frame has 13 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{deposition}{
a numeric vector
}
\item{LichenN}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex12.20)
}
\keyword{datasets}
