\name{ex12.24}
\alias{ex12.24}
\non_function{}
\title{data from exercise 12.24}
\description{
The \code{ex12.24} data frame has 6 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{SO.2dep.}{
a numeric vector
}
\item{Wt.loss}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex12.24)
}
\keyword{datasets}
