\name{ex12.37}
\alias{ex12.37}
\non_function{}
\title{data from exercise 12.37}
\description{
The \code{ex12.37} data frame has 10 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{pressure}{
a numeric vector
}
\item{time}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex12.37)
}
\keyword{datasets}
