\name{ex12.50}
\alias{ex12.50}
\non_function{}
\title{data from exercise 12.50}
\description{
The \code{ex12.50} data frame has 11 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{field}{
a numeric vector
}
\item{time}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex12.50)
}
\keyword{datasets}
