\name{ex12.55}
\alias{ex12.55}
\non_function{}
\title{data from exercise 12.55}
\description{
The \code{ex12.55} data frame has 12 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{age}{
a numeric vector
}
\item{damaged}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex12.55)
}
\keyword{datasets}
