\name{ex12.58}
\alias{ex12.58}
\non_function{}
\title{data from exercise 12.58}
\description{
The \code{ex12.58} data frame has 12 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{TOST}{
a numeric vector
}
\item{RBOT}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex12.58)
}
\keyword{datasets}
