\name{ex12.63}
\alias{ex12.63}
\non_function{}
\title{data from exercise 12.63}
\description{
The \code{ex12.63} data frame has 6 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{stiffnss}{
a numeric vector
}
\item{thicknss}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex12.63)
}
\keyword{datasets}
