\name{ex12.65}
\alias{ex12.65}
\non_function{}
\title{data from exercise 12.65}
\description{
The \code{ex12.65} data frame has 10 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{blood}{
a numeric vector
}
\item{gasoline}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex12.65)
}
\keyword{datasets}
