\name{ex12.68}
\alias{ex12.68}
\non_function{}
\title{data from exercise 12.68}
\description{
The \code{ex12.68} data frame has 8 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{RDF}{
a numeric vector
}
\item{eff}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex12.68)
}
\keyword{datasets}
