\name{ex12.69}
\alias{ex12.69}
\non_function{}
\title{data from exercise 12.69}
\description{
The \code{ex12.69} data frame has 13 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{drain.wt}{
a numeric vector
}
\item{Cl.trace}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex12.69)
}
\keyword{datasets}
