\name{ex12.71}
\alias{ex12.71}
\non_function{}
\title{data from exercise 12.71}
\description{
The \code{ex12.71} data frame has 13 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{austenite}{
a numeric vector
}
\item{wearLoss}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex12.71)
}
\keyword{datasets}
