\name{ex12.73}
\alias{ex12.73}
\non_function{}
\title{data from exercise 12.73}
\description{
The \code{ex12.73} data frame has 9 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{age}{
a numeric vector
}
\item{load}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex12.73)
}
\keyword{datasets}
