\name{ex12.75}
\alias{ex12.75}
\non_function{}
\title{data from exercise 12.75}
\description{
The \code{ex12.75} data frame has 9 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{absorb}{
a numeric vector
}
\item{peakVolt}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex12.75)
}
\keyword{datasets}
