\name{ex13.02}
\alias{ex13.02}
\non_function{}
\title{data from exercise 13.2}
\description{
The \code{ex13.02} data frame has 9 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{rate}{
a numeric vector
}
\item{stdResid}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.02)
}
\keyword{datasets}
