\name{ex13.04}
\alias{ex13.04}
\non_function{}
\title{data from exercise 13.4}
\description{
The \code{ex13.04} data frame has 10 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{thicknss}{
a numeric vector
}
\item{current}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.04)
}
\keyword{datasets}
