\name{ex13.06}
\alias{ex13.06}
\non_function{}
\title{data from exercise 13.6}
\description{
The \code{ex13.06} data frame has 6 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{density}{
a numeric vector
}
\item{moisture}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.06)
}
\keyword{datasets}
