\name{ex13.08}
\alias{ex13.08}
\non_function{}
\title{data from exercise 13.8}
\description{
The \code{ex13.08} data frame has 10 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{age}{
a numeric vector
}
\item{growth}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.08)
}
\keyword{datasets}
