\name{ex13.09}
\alias{ex13.09}
\non_function{}
\title{data from exercise 13.9}
\description{
The \code{ex13.09} data frame has 44 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{x}{
a numeric vector
}
\item{y}{
a numeric vector
}
\item{set}{
a factor with levels
\code{a} 
\code{b} 
\code{c} 
\code{d} 
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.09)
}
\keyword{datasets}
