\name{ex13.14}
\alias{ex13.14}
\non_function{}
\title{data from exercise 13.14}
\description{
The \code{ex13.14} data frame has 14 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{weight}{
a numeric vector
}
\item{clearnce}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.14)
}
\keyword{datasets}
