\name{ex13.16}
\alias{ex13.16}
\non_function{}
\title{data from exercise 13.16}
\description{
The \code{ex13.16} data frame has 12 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Spectral}{
a numeric vector
}
\item{ln.L178.}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.16)
}
\keyword{datasets}
