\name{ex13.17}
\alias{ex13.17}
\non_function{}
\title{data from exercise 13.17}
\description{
The \code{ex13.17} data frame has 13 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{MassRate}{
a numeric vector
}
\item{FlameLen}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.17)
}
\keyword{datasets}
