\name{ex13.18}
\alias{ex13.18}
\non_function{}
\title{data from exercise 13.18}
\description{
The \code{ex13.18} data frame has 4 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Conc.}{
a numeric vector
}
\item{pH}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.18)
}
\keyword{datasets}
