\name{ex13.19}
\alias{ex13.19}
\non_function{}
\title{data from exercise 13.19}
\description{
The \code{ex13.19} data frame has 18 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Temp}{
a numeric vector
}
\item{Lifetime}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.19)
}
\keyword{datasets}
