\name{ex13.21}
\alias{ex13.21}
\non_function{}
\title{data from exercise 13.21}
\description{
The \code{ex13.21} data frame has 8 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{thicknss}{
a numeric vector
}
\item{conduct}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.21)
}
\keyword{datasets}
