\name{ex13.24}
\alias{ex13.24}
\non_function{}
\title{data from exercise 13.24}
\description{
The \code{ex13.24} data frame has 40 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{age}{
a numeric vector
}
\item{Kyphosis}{
a factor with levels
\code{N} 
\code{Y} 
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.24)
}
\keyword{datasets}
