\name{ex13.25}
\alias{ex13.25}
\non_function{}
\title{data from exercise 13.25}
\description{
The \code{ex13.25} data frame has 14 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Success}{
a numeric vector
}
\item{Failure}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.25)
}
\keyword{datasets}
