\name{ex13.27}
\alias{ex13.27}
\non_function{}
\title{data from exercise 13.27}
\description{
The \code{ex13.27} data frame has 8 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{time}{
a numeric vector
}
\item{conc}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.27)
}
\keyword{datasets}
