\name{ex13.29}
\alias{ex13.29}
\non_function{}
\title{data from exercise 13.29}
\description{
The \code{ex13.29} data frame has 5 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{x}{
a numeric vector
}
\item{y}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.29)
}
\keyword{datasets}
