\name{ex13.30}
\alias{ex13.30}
\non_function{}
\title{data from exercise 13.30}
\description{
The \code{ex13.30} data frame has 14 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{water}{
a numeric vector
}
\item{yield}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.30)
}
\keyword{datasets}
