\name{ex13.31}
\alias{ex13.31}
\non_function{}
\title{data from exercise 13.31}
\description{
The \code{ex13.31} data frame has 7 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{velocity}{
a numeric vector
}
\item{conversn}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.31)
}
\keyword{datasets}
