\name{ex13.32}
\alias{ex13.32}
\non_function{}
\title{data from exercise 13.32}
\description{
The \code{ex13.32} data frame has 16 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{soil.Ph}{
a numeric vector
}
\item{conc}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.32)
}
\keyword{datasets}
