\name{ex13.33}
\alias{ex13.33}
\non_function{}
\title{data from exercise 13.33}
\description{
The \code{ex13.33} data frame has 7 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{angle}{
a numeric vector
}
\item{efficncy}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.33)
}
\keyword{datasets}
