\name{ex13.34}
\alias{ex13.34}
\non_function{}
\title{data from exercise 13.34}
\description{
The \code{ex13.34} data frame has 13 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{cations}{
a numeric vector
}
\item{exchange}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.34)
}
\keyword{datasets}
