\name{ex13.35}
\alias{ex13.35}
\non_function{}
\title{data from exercise 13.35}
\description{
The \code{ex13.35} data frame has 5 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{tempture}{
a numeric vector
}
\item{rate}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.35)
}
\keyword{datasets}
