\name{ex13.40}
\alias{ex13.40}
\non_function{}
\title{data from exercise 13.40}
\description{
The \code{ex13.40} data frame has 14 rows and 4 columns.
}
\format{
This data frame contains the following columns:
}
\arguments{
\item{-1}{
a numeric vector
}
\item{-1}{
a numeric vector
}
\item{0}{
a numeric vector
}
\item{18.3}{
a numeric vector
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(ex13.40)
}
\keyword(datasets)
