\name{ex13.42}
\alias{ex13.42}
\non_function{}
\title{data from exercise 13.42}
\description{
The \code{ex13.42} data frame has 11 rows and 3 columns.
}
\format{
This data frame contains the following columns:
}
\arguments{
\item{16.7}{
a numeric vector
}
\item{30}{
a numeric vector
}
\item{210}{
a numeric vector
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(ex13.42)
}
\keyword(datasets)
