\name{ex13.44}
\alias{ex13.44}
\non_function{}
\title{data from exercise 13.44}
\description{
The \code{ex13.44} data frame has 13 rows and 3 columns.
}
\format{
This data frame contains the following columns:
}
\arguments{
\item{14.7}{
a numeric vector
}
\item{8.9}{
a numeric vector
}
\item{31.5}{
a numeric vector
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(ex13.44)
}
\keyword(datasets)
