\name{ex13.46}
\alias{ex13.46}
\non_function{}
\title{data from exercise 13.46}
\description{
The \code{ex13.46} data frame has 30 rows and 5 columns.
}
\format{
This data frame contains the following columns:
}
\arguments{
\item{-1}{
a numeric vector
}
\item{-1}{
a numeric vector
}
\item{-1}{
a numeric vector
}
\item{-1}{
a numeric vector
}
\item{83.9}{
a numeric vector
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.46)
}
\keyword(datasets)
