\name{ex13.47}
\alias{ex13.47}
\non_function{}
\title{data from exercise 13.47}
\description{
The \code{ex13.47} data frame has 30 rows and 6 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Row}{
a numeric vector
}
\item{Plastics}{
a numeric vector
}
\item{Paper}{
a numeric vector
}
\item{Garbage}{
a numeric vector
}
\item{Water}{
a numeric vector
}
\item{Energy.content}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.47)
}
\keyword{datasets}
