\name{ex13.48}
\alias{ex13.48}
\non_function{}
\title{data from exercise 13.48}
\description{
The \code{ex13.48} data frame has 15 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{x1}{
a numeric vector
}
\item{x2}{
a numeric vector
}
\item{x3}{
a numeric vector
}
\item{y}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.48)
}
\keyword{datasets}
