\name{ex13.50}
\alias{ex13.50}
\non_function{}
\title{data from exercise 13.50}
\description{
The \code{ex13.50} data frame has 14 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Col1}{
a numeric vector
}
\item{Col2}{
a numeric vector
}
\item{Col3}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.50)
}
\keyword{datasets}
