\name{ex13.51}
\alias{ex13.51}
\non_function{}
\title{data from exercise 13.51}
\description{
The \code{ex13.51} data frame has 14 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{shear}{
a numeric vector
}
\item{depth}{
a numeric vector
}
\item{water}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.51)
}
\keyword{datasets}
