\name{ex13.52}
\alias{ex13.52}
\non_function{}
\title{data from exercise 13.52}
\description{
The \code{ex13.52} data frame has 31 rows and 5 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Bright}{
a numeric vector
}
\item{H2O2}{
a numeric vector
}
\item{NaOH}{
a numeric vector
}
\item{Silicate}{
a numeric vector
}
\item{Tempture}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.52)
}
\keyword{datasets}
