\name{ex13.53}
\alias{ex13.53}
\non_function{}
\title{data from exercise 13.53}
\description{
The \code{ex13.53} data frame has 10 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{q}{
a numeric vector
}
\item{a}{
a numeric vector
}
\item{b}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.53)
}
\keyword{datasets}
