\name{ex13.64}
\alias{ex13.64}
\non_function{}
\title{data from exercise 13.64}
\description{
The \code{ex13.64} data frame has 16 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Log.edges.}{
a numeric vector
}
\item{Log.time.}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.64)
}
\keyword{datasets}
