\name{ex13.65}
\alias{ex13.65}
\non_function{}
\title{data from exercise 13.65}
\description{
The \code{ex13.65} data frame has 18 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Pressure}{
a numeric vector
}
\item{Temperature}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.65)
}
\keyword{datasets}
