\name{ex13.67}
\alias{ex13.67}
\non_function{}
\title{data from exercise 13.67}
\description{
The \code{ex13.67} data frame has 32 rows and 7 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Obs}{
a numeric vector
}
\item{pdconc}{
a numeric vector
}
\item{niconc}{
a numeric vector
}
\item{pH}{
a numeric vector
}
\item{temp}{
a numeric vector
}
\item{currdens}{
a numeric vector
}
\item{pallcont}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.67)
}
\keyword{datasets}
