\name{ex13.69}
\alias{ex13.69}
\non_function{}
\title{data from exercise 13.69}
\description{
The \code{ex13.69} data frame has 8 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{power}{
a numeric vector
}
\item{freq}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.69)
}
\keyword{datasets}
