\name{ex13.70}
\alias{ex13.70}
\non_function{}
\title{data from exercise 13.70}
\description{
The \code{ex13.70} data frame has 12 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{log.con.}{
a numeric vector
}
\item{Li2O}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.70)
}
\keyword{datasets}
