\name{ex13.71}
\alias{ex13.71}
\non_function{}
\title{data from exercise 13.71}
\description{
The \code{ex13.71} data frame has 10 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{height}{
a numeric vector
}
\item{log.Mn.}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.71)
}
\keyword{datasets}
