\name{ex13.72}
\alias{ex13.72}
\non_function{}
\title{data from exercise 13.72}
\description{
The \code{ex13.72} data frame has 9 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{x1}{
a numeric vector
}
\item{x2}{
a numeric vector
}
\item{y}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex13.72)
}
\keyword{datasets}
