\name{ex14.12}
\alias{ex14.12}
\non_function{}
\title{data from exercise 14.12}
\description{
The \code{ex14.12} data frame has 4 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{male.children}{
a numeric vector
}
\item{Frequency}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex14.12)
}
\keyword{datasets}
