\name{ex14.13}
\alias{ex14.13}
\non_function{}
\title{data from exercise 14.13}
\description{
The \code{ex14.13} data frame has 3 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{ovaries.developed}{
a numeric vector
}
\item{Observed.count}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex14.13)
}
\keyword{datasets}
