\name{ex14.14}
\alias{ex14.14}
\non_function{}
\title{data from exercise 14.14}
\description{
The \code{ex14.14} data frame has 12 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{x}{
a numeric vector
}
\item{observed}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex14.14)
}
\keyword{datasets}
