\name{ex14.15}
\alias{ex14.15}
\non_function{}
\title{data from exercise 14.15}
\description{
The \code{ex14.15} data frame has 5 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Number.defective}{
a numeric vector
}
\item{Frequency}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex14.15)
}
\keyword{datasets}
