\name{ex14.16}
\alias{ex14.16}
\non_function{}
\title{data from exercise 14.16}
\description{
The \code{ex14.16} data frame has 10 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Number.exchanges}{
a numeric vector
}
\item{Observed.counts}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex14.16)
}
\keyword{datasets}
