\name{ex14.17}
\alias{ex14.17}
\non_function{}
\title{data from exercise 14.17}
\description{
The \code{ex14.17} data frame has 13 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{borers}{
a numeric vector
}
\item{freq}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex14.17)
}
\keyword{datasets}
