\name{ex14.18}
\alias{ex14.18}
\non_function{}
\title{data from exercise 14.18}
\description{
The \code{ex14.18} data frame has 5 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Rate..per.day.}{
a factor with levels
\code{.100-below .150} 
\code{.150-below .200} 
\code{.200-below .250} 
\code{.250 or more} 
\code{Below .100} 
}
\item{Frequency}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex14.18)
}
\keyword{datasets}
