\name{ex14.29}
\alias{ex14.29}
\non_function{}
\title{data from exercise 14.29}
\description{
The \code{ex14.29} data frame has 6 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{M.M}{
a numeric vector
}
\item{M.F}{
a numeric vector
}
\item{F.F}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex14.29)
}
\keyword{datasets}
