\name{ex14.31}
\alias{ex14.31}
\non_function{}
\title{data from exercise 14.31}
\description{
The \code{ex14.31} data frame has 12 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{count}{
a numeric vector
}
\item{Size}{
a factor with levels
\code{Compact} 
\code{Fullsize} 
\code{Midsize} 
\code{Subcompact} 
}
\item{dist}{
a factor with levels
\code{0-<10} 
\code{10-<20} 
\code{>=20} 
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex14.31)
}
\keyword{datasets}
