\name{ex14.32}
\alias{ex14.32}
\non_function{}
\title{data from exercise 14.32}
\description{
The \code{ex14.32} data frame has 3 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Liberal}{
a numeric vector
}
\item{Consrvtv}{
a numeric vector
}
\item{Other}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex14.32)
}
\keyword{datasets}
