\name{ex14.38}
\alias{ex14.38}
\non_function{}
\title{data from exercise 14.38}
\description{
The \code{ex14.38} data frame has 3 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Parsitiz}{
a numeric vector
}
\item{Nonparas}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex14.38)
}
\keyword{datasets}
