\name{ex14.40}
\alias{ex14.40}
\non_function{}
\title{data from exercise 14.40}
\description{
The \code{ex14.40} data frame has 4 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Sport}{
a factor with levels
\code{Baseball} 
\code{Basketball} 
\code{Football} 
\code{Hockey} 
}
\item{Leader.Wins}{
a numeric vector
}
\item{Leader.Loses}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex14.40)
}
\keyword{datasets}
