\name{ex14.41}
\alias{ex14.41}
\non_function{}
\title{data from exercise 14.41}
\description{
The \code{ex14.41} data frame has 3 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Never}{
a numeric vector
}
\item{Occasion}{
a numeric vector
}
\item{Regular}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex14.41)
}
\keyword{datasets}
