\name{ex14.44}
\alias{ex14.44}
\non_function{}
\title{data from exercise 14.44}
\description{
The \code{ex14.44} data frame has 5 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{sample}{
a numeric vector
}
\item{item.prc}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex14.44)
}
\keyword{datasets}
