\name{ex15.03}
\alias{ex15.03}
\non_function{}
\title{data from exercise 15.3}
\description{
The \code{ex15.03} data frame has 14 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{pH}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex15.03)
}
\keyword{datasets}
