\name{ex15.08}
\alias{ex15.08}
\non_function{}
\title{data from exercise 15.8}
\description{
The \code{ex15.08} data frame has 25 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{toughnss}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex15.08)
}
\keyword{datasets}
