\name{ex15.10}
\alias{ex15.10}
\non_function{}
\title{data from exercise 15.10}
\description{
The \code{ex15.10} data frame has 5 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{adhesv.1}{
a numeric vector
}
\item{adhesv.2}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex15.10)
}
\keyword{datasets}
