\name{ex15.11}
\alias{ex15.11}
\non_function{}
\title{data from exercise 15.11}
\description{
The \code{ex15.11} data frame has 14 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Time}{
a numeric vector
}
\item{wood}{
a factor with levels
\code{Oak} 
\code{Pine} 
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex15.11)
}
\keyword{datasets}
