\name{ex15.12}
\alias{ex15.12}
\non_function{}
\title{data from exercise 15.12}
\description{
The \code{ex15.12} data frame has 8 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Original}{
a numeric vector
}
\item{Modified}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex15.12)
}
\keyword{datasets}
