\name{ex15.13}
\alias{ex15.13}
\non_function{}
\title{data from exercise 15.13}
\description{
The \code{ex15.13} data frame has 10 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Orange.J}{
a numeric vector
}
\item{Ascorbic}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex15.13)
}
\keyword{datasets}
