\name{ex15.23}
\alias{ex15.23}
\non_function{}
\title{data from exercise 15.23}
\description{
The \code{ex15.23} data frame has 5 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Region.1}{
a numeric vector
}
\item{Region.2}{
a numeric vector
}
\item{Region.3}{
a numeric vector
}
\item{Region.4}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex15.23)
}
\keyword{datasets}
