\name{ex15.24}
\alias{ex15.24}
\non_function{}
\title{data from exercise 15.24}
\description{
The \code{ex15.24} data frame has 9 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{X1}{
a numeric vector
}
\item{X2}{
a numeric vector
}
\item{X3}{
a numeric vector
}
\item{X4}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex15.24)
}
\keyword{datasets}
