\name{ex15.25}
\alias{ex15.25}
\non_function{}
\title{data from exercise 15.25}
\description{
The \code{ex15.25} data frame has 22 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{cortisol}{
a numeric vector
}
\item{Group}{
a factor with levels
\code{A} 
\code{B} 
\code{C} 
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex15.25)
}
\keyword{datasets}
