\name{ex15.27}
\alias{ex15.27}
\non_function{}
\title{data from exercise 15.27}
\description{
The \code{ex15.27} data frame has 10 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Dog}{
a numeric vector
}
\item{Isoflurane}{
a numeric vector
}
\item{Halothane}{
a numeric vector
}
\item{Cyclopropane}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex15.27)
}
\keyword{datasets}
