\name{ex15.28}
\alias{ex15.28}
\non_function{}
\title{data from exercise 15.28}
\description{
The \code{ex15.28} data frame has 8 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Subject}{
a numeric vector
}
\item{Potato}{
a numeric vector
}
\item{Rice}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex15.28)
}
\keyword{datasets}
