\name{ex15.30}
\alias{ex15.30}
\non_function{}
\title{data from exercise 15.30}
\description{
The \code{ex15.30} data frame has 5 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Treatment.I}{
a numeric vector
}
\item{Treatment.II}{
a numeric vector
}
\item{Treatment.III}{
a numeric vector
}
\item{Treatment.IV}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex15.30)
}
\keyword{datasets}
