\name{ex15.32}
\alias{ex15.32}
\non_function{}
\title{data from exercise 15.32}
\description{
The \code{ex15.32} data frame has 13 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Time}{
a numeric vector
}
\item{Gait}{
a factor with levels
\code{Diagonal} 
\code{Lateral} 
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex15.32)
}
\keyword{datasets}
