\name{ex15.35}
\alias{ex15.35}
\non_function{}
\title{data from exercise 15.35}
\description{
The \code{ex15.35} data frame has 9 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{thickness}{
a numeric vector
}
\item{condition}{
a factor with levels
\code{Control} 
\code{SIDS} 
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex15.35)
}
\keyword{datasets}
