\name{ex16.05}
\alias{ex16.05}
\non_function{}
\title{data from exercise 16.5}
\description{
The \code{ex16.05} data frame has 22 rows and 5 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{moist1}{
a numeric vector
}
\item{moist2}{
a numeric vector
}
\item{moist3}{
a numeric vector
}
\item{moist4}{
a numeric vector
}
\item{moist5}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex16.05)
}
\keyword{datasets}
