\name{ex16.06}
\alias{ex16.06}
\non_function{}
\title{data from exercise 16.6}
\description{
The \code{ex16.06} data frame has 22 rows and 5 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Col1}{
a numeric vector
}
\item{Col2}{
a numeric vector
}
\item{Col3}{
a numeric vector
}
\item{Col4}{
a numeric vector
}
\item{Col5}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex16.06)
}
\keyword{datasets}
