\name{ex16.09}
\alias{ex16.09}
\non_function{}
\title{data from exercise 16.9}
\description{
The \code{ex16.09} data frame has 24 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{xbar}{
a numeric vector
}
\item{stderr}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex16.09)
}
\keyword{datasets}
