\name{ex16.14}
\alias{ex16.14}
\non_function{}
\title{data from exercise 16.14}
\description{
The \code{ex16.14} data frame has 24 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{C1}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex16.14)
}
\keyword{datasets}
