\name{ex16.41}
\alias{ex16.41}
\non_function{}
\title{data from exercise 16.41}
\description{
The \code{ex16.41} data frame has 22 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{tensile1}{
a numeric vector
}
\item{tensile2}{
a numeric vector
}
\item{tensile3}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex16.41)
}
\keyword{datasets}
