\name{ex16.43}
\alias{ex16.43}
\non_function{}
\title{data from exercise 16.43}
\description{
The \code{ex16.43} data frame has 20 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{n}{
a numeric vector
}
\item{xbar}{
a numeric vector
}
\item{stderr}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(ex16.43)
}
\keyword{datasets}
