\name{xmp01.16}
\alias{xmp01.16}
\non_function{}
\title{data from Example 1.16}
\description{
The \code{xmp01.16} data frame has 11 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Strength}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp01.16)
}
\keyword{datasets}
