\name{xmp01.19}
\alias{xmp01.19}
\non_function{}
\title{data from Example 1.19}
\description{
The \code{xmp01.19} data frame has 25 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{C1}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp01.19)
}
\keyword{datasets}
