\name{xmp06.03}
\alias{xmp06.03}
\non_function{}
\title{data from Example 6.3}
\description{
The \code{xmp06.03} data frame has 8 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{C1}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp06.03)
}
\keyword{datasets}
