\name{xmp06.13}
\alias{xmp06.13}
\non_function{}
\title{data from Example 6.13}
\description{
The \code{xmp06.13} data frame has 420 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{goals}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp06.13)
}
\keyword{datasets}
