\name{xmp07.11}
\alias{xmp07.11}
\non_function{}
\title{data from Example 7.11}
\description{
The \code{xmp07.11} data frame has 16 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{C1}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp07.11)
}
\keyword{datasets}
