\name{xmp07.15}
\alias{xmp07.15}
\non_function{}
\title{data from Example 7.15}
\description{
The \code{xmp07.15} data frame has 17 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{voltage}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp07.15)
}
\keyword{datasets}
