\name{xmp08.08}
\alias{xmp08.08}
\non_function{}
\title{data from Example 8.8}
\description{
The \code{xmp08.08} data frame has 52 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{C1}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp08.08)
}
\keyword{datasets}
