\name{xmp09.10}
\alias{xmp09.10}
\non_function{}
\title{data from Example 9.10}
\description{
The \code{xmp09.10} data frame has 16 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Observation}{
a numeric vector
}
\item{1.min}{
a numeric vector
}
\item{4.weeks}{
a numeric vector
}
\item{Difference}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp09.10)
}
\keyword{datasets}
