\name{xmp10.01}
\alias{xmp10.01}
\non_function{}
\title{data from Example 10.1}
\description{
The \code{xmp10.01} data frame has 24 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{strength}{
a numeric vector
}
\item{type}{
a factor with levels
\code{A} 
\code{B} 
\code{C} 
\code{D} 
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp10.01)
}
\keyword{datasets}
