\name{xmp10.03}
\alias{xmp10.03}
\non_function{}
\title{data from Example 10.3}
\description{
The \code{xmp10.03} data frame has 15 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Soiling}{
a numeric vector
}
\item{Mixture}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp10.03)
}
\keyword{datasets}
