\name{xmp10.05}
\alias{xmp10.05}
\non_function{}
\title{data from Example 10.5}
\description{
The \code{xmp10.05} data frame has 20 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{time}{
a numeric vector
}
\item{gm.kg}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp10.05)
}
\keyword{datasets}
