\name{xmp10.10}
\alias{xmp10.10}
\non_function{}
\title{data from Example 10.10}
\description{
The \code{xmp10.10} data frame has 18 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Time.ns.}{
a numeric vector
}
\item{Rail}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp10.10)
}
\keyword{datasets}
