\name{xmp11.01}
\alias{xmp11.01}
\non_function{}
\title{data from Example 11.1}
\description{
The \code{xmp11.01} data frame has 12 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{strength}{
a numeric vector
}
\item{brand}{
a numeric vector
}
\item{treatment}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp11.01)
}
\keyword{datasets}
