\name{xmp11.05}
\alias{xmp11.05}
\non_function{}
\title{data from Example 11.5}
\description{
The \code{xmp11.05} data frame has 20 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{power}{
a numeric vector
}
\item{humid}{
a numeric vector
}
\item{brand}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp11.05)
}
\keyword{datasets}
