\name{xmp11.06}
\alias{xmp11.06}
\non_function{}
\title{data from Example 11.6}
\description{
The \code{xmp11.06} data frame has 24 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Resp}{
a numeric vector
}
\item{Stimulus}{
a numeric vector
}
\item{Subject}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp11.06)
}
\keyword{datasets}
