\name{xmp11.07}
\alias{xmp11.07}
\non_function{}
\title{data from Example 11.7}
\description{
The \code{xmp11.07} data frame has 36 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Yield}{
a numeric vector
}
\item{Variety}{
a numeric vector
}
\item{Density}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp11.07)
}
\keyword{datasets}
