\name{xmp11.11}
\alias{xmp11.11}
\non_function{}
\title{data from Example 11.11}
\description{
The \code{xmp11.11} data frame has 96 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Tempr}{
a numeric vector
}
\item{Period}{
a numeric vector
}
\item{Strain}{
a numeric vector
}
\item{Coat}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp11.11)
}
\keyword{datasets}
