\name{xmp11.12}
\alias{xmp11.12}
\non_function{}
\title{data from Example 11.12}
\description{
The \code{xmp11.12} data frame has 36 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{abrasion}{
a numeric vector
}
\item{row}{
a numeric vector
}
\item{column}{
a numeric vector
}
\item{humidity}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp11.12)
}
\keyword{datasets}
