\name{xmp11.16}
\alias{xmp11.16}
\non_function{}
\title{data from Example 11.16}
\description{
The \code{xmp11.16} data frame has 16 rows and 6 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{strength}{
a numeric vector
}
\item{tempture}{
a numeric vector
}
\item{gas}{
a numeric vector
}
\item{conc}{
a numeric vector
}
\item{block}{
a numeric vector
}
\item{c.1...1...1..1..1...1...1..1..1...1...1..1..1...1...1..1.}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp11.16)
}
\keyword{datasets}
