\name{xmp12.01}
\alias{xmp12.01}
\non_function{}
\title{data from Example 12.1}
\description{
The \code{xmp12.01} data frame has 30 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Obs}{
a numeric vector
}
\item{x}{
a numeric vector
}
\item{y}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp12.01)
}
\keyword{datasets}
