\name{xmp12.04}
\alias{xmp12.04}
\non_function{}
\title{data from Example 12.4}
\description{
The \code{xmp12.04} data frame has 15 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{x}{
a numeric vector
}
\item{y}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp12.04)
}
\keyword{datasets}
