\name{xmp12.06}
\alias{xmp12.06}
\non_function{}
\title{data from Example 12.6}
\description{
The \code{xmp12.06} data frame has 11 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{traffic}{
a numeric vector
}
\item{lead}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp12.06)
}
\keyword{datasets}
