\name{xmp12.08}
\alias{xmp12.08}
\non_function{}
\title{data from Example 12.8}
\description{
The \code{xmp12.08} data frame has 14 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{strength}{
a numeric vector
}
\item{attenuat}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp12.08)
}
\keyword{datasets}
