\name{xmp12.11}
\alias{xmp12.11}
\non_function{}
\title{data from Example 12.11}
\description{
The \code{xmp12.11} data frame has 20 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{x}{
a numeric vector
}
\item{y}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp12.11)
}
\keyword{datasets}
