\name{xmp12.15}
\alias{xmp12.15}
\non_function{}
\title{data from Example 12.15}
\description{
The \code{xmp12.15} data frame has 16 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{ozone}{
a numeric vector
}
\item{carbon}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp12.15)
}
\keyword{datasets}
