\name{xmp13.01}
\alias{xmp13.01}
\non_function{}
\title{data from Example 13.1}
\description{
The \code{xmp13.01} data frame has 14 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{rate}{
a numeric vector
}
\item{emission}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp13.01)
}
\keyword{datasets}
