\name{xmp13.03}
\alias{xmp13.03}
\non_function{}
\title{data from Example 13.3}
\description{
The \code{xmp13.03} data frame has 12 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{time}{
a numeric vector
}
\item{ToolLife}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp13.03)
}
\keyword{datasets}
