\name{xmp13.04}
\alias{xmp13.04}
\non_function{}
\title{data from Example 13.4}
\description{
The \code{xmp13.04} data frame has 11 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{time}{
a numeric vector
}
\item{Ethylene}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp13.04)
}
\keyword{datasets}
