\name{xmp13.09}
\alias{xmp13.09}
\non_function{}
\title{data from Example 13.9}
\description{
The \code{xmp13.09} data frame has 8 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{tempture}{
a numeric vector
}
\item{strength}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp13.09)
}
\keyword{datasets}
