\name{xmp13.11}
\alias{xmp13.11}
\non_function{}
\title{data from Example 13.11}
\description{
The \code{xmp13.11} data frame has 30 rows and 6 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Observation}{
a numeric vector
}
\item{Force}{
a numeric vector
}
\item{Power}{
a numeric vector
}
\item{Temperature}{
a numeric vector
}
\item{Time}{
a numeric vector
}
\item{Strength}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp13.11)
}
\keyword{datasets}
