\name{xmp13.14}
\alias{xmp13.14}
\non_function{}
\title{data from Example 13.14}
\description{
The \code{xmp13.14} data frame has 13 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{index}{
a numeric vector
}
\item{iron}{
a numeric vector
}
\item{aluminum}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp13.14)
}
\keyword{datasets}
