\name{xmp13.15}
\alias{xmp13.15}
\non_function{}
\title{data from Example 13.15}
\description{
The \code{xmp13.15} data frame has 30 rows and 5 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{press}{
a numeric vector
}
\item{HCHO}{
a numeric vector
}
\item{catalyst}{
a numeric vector
}
\item{temp}{
a numeric vector
}
\item{time}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp13.15)
}
\keyword{datasets}
