\name{xmp13.17}
\alias{xmp13.17}
\non_function{}
\title{data from Example 13.17}
\description{
The \code{xmp13.17} data frame has 27 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{life}{
a numeric vector
}
\item{speed}{
a numeric vector
}
\item{load}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp13.17)
}
\keyword{datasets}
