\name{xmp13.18}
\alias{xmp13.18}
\non_function{}
\title{data from Example 13.18}
\description{
The \code{xmp13.18} data frame has 31 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{tar}{
a numeric vector
}
\item{speed}{
a numeric vector
}
\item{tempture}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp13.18)
}
\keyword{datasets}
