\name{xmp14.10}
\alias{xmp14.10}
\non_function{}
\title{data from Example 14.10}
\description{
The \code{xmp14.10} data frame has 49 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Cholstrl}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp14.10)
}
\keyword{datasets}
