\name{xmp14.13}
\alias{xmp14.13}
\non_function{}
\title{data from Example 14.13}
\description{
The \code{xmp14.13} data frame has 3 rows and 6 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Production.Line}{
a numeric vector
}
\item{Blemish}{
a numeric vector
}
\item{Crack}{
a numeric vector
}
\item{Location}{
a numeric vector
}
\item{Missing}{
a numeric vector
}
\item{Other}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp14.13)
}
\keyword{datasets}
