\name{xmp14.14}
\alias{xmp14.14}
\non_function{}
\title{data from Example 14.14}
\description{
The \code{xmp14.14} data frame has 3 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Substand}{
a numeric vector
}
\item{Standard}{
a numeric vector
}
\item{Modern}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp14.14)
}
\keyword{datasets}
