\name{xmp15.01}
\alias{xmp15.01}
\non_function{}
\title{data from Example 15.1}
\description{
The \code{xmp15.01} data frame has 15 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{C1}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp15.01)
}
\keyword{datasets}
