\name{xmp15.02}
\alias{xmp15.02}
\non_function{}
\title{data from Example 15.2}
\description{
The \code{xmp15.02} data frame has 8 rows and 5 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Log}{
a numeric vector
}
\item{Solvent.1}{
a numeric vector
}
\item{Solvent.2}{
a numeric vector
}
\item{Difference}{
a numeric vector
}
\item{Signed.rank}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp15.02)
}
\keyword{datasets}
