\name{xmp15.03}
\alias{xmp15.03}
\non_function{}
\title{data from Example 15.3}
\description{
The \code{xmp15.03} data frame has 25 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{xi}{
a numeric vector
}
\item{Signed.Rank}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp15.03)
}
\keyword{datasets}
