\name{xmp15.04}
\alias{xmp15.04}
\non_function{}
\title{data from Example 15.4}
\description{
The \code{xmp15.04} data frame has 12 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{conc}{
a numeric vector
}
\item{Area}{
a factor with levels
\code{Polluted} 
\code{Unpolluted} 
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp15.04)
}
\keyword{datasets}
