\name{xmp15.09}
\alias{xmp15.09}
\non_function{}
\title{data from Example 15.9}
\description{
The \code{xmp15.09} data frame has 7 rows and 5 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{X4.inch}{
a numeric vector
}
\item{X6.inch}{
a numeric vector
}
\item{X8.inch}{
a numeric vector
}
\item{X10.inch}{
a numeric vector
}
\item{X12.inch}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp15.09)
}
\keyword{datasets}
