\name{xmp15.10}
\alias{xmp15.10}
\non_function{}
\title{data from Example 15.10}
\description{
The \code{xmp15.10} data frame has 32 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{potental}{
a numeric vector
}
\item{emotion}{
a numeric vector
}
\item{subject}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp15.10)
}
\keyword{datasets}
