\name{xmp16.01}
\alias{xmp16.01}
\non_function{}
\title{data from Example 16.1}
\description{
The \code{xmp16.01} data frame has 25 rows and 3 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{visc1}{
a numeric vector
}
\item{visc2}{
a numeric vector
}
\item{visc3}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp16.01)
}
\keyword{datasets}
