\name{xmp16.06}
\alias{xmp16.06}
\non_function{}
\title{data from Example 16.6}
\description{
The \code{xmp16.06} data frame has 25 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{defects}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp16.06)
}
\keyword{datasets}
