\name{xmp16.07}
\alias{xmp16.07}
\non_function{}
\title{data from Example 16.7}
\description{
The \code{xmp16.07} data frame has 24 rows and 1 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{flaws}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp16.07)
}
\keyword{datasets}
