\name{xmp16.08}
\alias{xmp16.08}
\non_function{}
\title{data from Example 16.8}
\description{
The \code{xmp16.08} data frame has 16 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{obs1}{
a numeric vector
}
\item{obs2}{
a numeric vector
}
\item{obs3}{
a numeric vector
}
\item{obs4}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
data(xmp16.08)
}
\keyword{datasets}
