\name{ex09.11}
\alias{ex09.11}
\docType{data}
\title{data from exercise 9.11}
\description{
The \code{ex09.11} data frame has 2 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Age}{
a numeric vector
}
\item{n}{
a numeric vector
}
\item{Mean}{
a numeric vector
}
\item{StdDev}{
a numeric vector
}
}
}
%\details{}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering and the Sciences (6th ed)}, Duxbury
}
\examples{
data(ex09.12)
}
\keyword{datasets}
