\name{ex10.37}
\alias{ex10.37}
\docType{data}
\title{data from exercise 10.37}
\description{
  Motor vibration for 5 different brands of motor bearing
}
\usage{data(ex10.37)}
\format{
  A data frame with 30 observations on the following 2 variables.
  \describe{
    \item{vibration}{a numeric vector}
    \item{Brand}{a factor with levels \code{1} \code{2} \code{3} \code{4} \code{5}}
  }
}
%\details{}
\source{
  Devore, J. L. (2003) \emph{Probability and Statistics for Engineering
    and the Sciences (6th ed)}, Duxbury
}
\references{
  ``Increasing Market Share Through Improved Product and Process Design:
  An Experimental Approach'', \emph{Quality Engineering}, 1991: 361-369.
}
\examples{
data(ex10.37)
}
\keyword{datasets}
