\name{xmp09.09}
\alias{xmp09.09}
\non_function{}
\title{data from Example 9.9}
\description{
The \code{xmp09.09} data frame has 16 rows and 4 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Subject}{
a numeric vector
}
\item{Before}{
a numeric vector
}
\item{After}{
a numeric vector
}
\item{Difference}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp09.09)
}
\keyword{datasets}
