\name{xmp13.05}
\alias{xmp13.05}
\non_function{}
\title{data from Example 13.5}
\description{
The \code{xmp13.05} data frame has 24 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{Temperature}{
a numeric vector
}
\item{Failure}{
a factor with levels
\code{N} 
\code{Y} 
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp13.05)
}
\keyword{datasets}
