\name{xmp13.06}
\alias{xmp13.06}
\non_function{}
\title{data from Example 13.6}
\description{
The \code{xmp13.06} data frame has 16 rows and 2 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{days}{
a numeric vector
}
\item{yield}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp13.06)
}
\keyword{datasets}
