\name{xmp13.12}
\alias{xmp13.12}
\non_function{}
\title{data from Example 13.12}
\description{
The \code{xmp13.12} data frame has 9 rows and 5 columns.
}
\format{
This data frame contains the following columns:
\describe{
\item{x1}{
a numeric vector
}
\item{x2}{
a numeric vector
}
\item{x1x2}{
a numeric vector
}
\item{28-day.compt.str..(Mpa)}{
a numeric vector
}
\item{Absorbability.(\%)}{
a numeric vector
}
}
}
\details{
}
\source{
  Devore, J. L. (2000) \emph{Probability and Statistics for Engineering and the Sciences (5th ed)}, Duxbury
}
\examples{
library(Devore5)
data(xmp13.12)
}
\keyword{datasets}
